requirements.txt
streamlit==1.33.0
openai-whisper==20231117